changed the way we handle victories and ties.
when a player has won it will set the boolean in the player (hasWon) to
true.

in playGame. at the end of each player's turn will check if the player's
won flag or tie flag is set to true.

will then display message and system.exit

this might help with the minimax